Cart quantitites stack

+++ front/js/product.js
@@ -4,18 +4,18 @@ console.log(productId);
 
 fetch(`http://localhost:3000/api/products/${productId}`)
   .then((res) => res.json())
-  .then((product) => productData(product))
+  .then((product) => sofa(product))
   .catch((err) => {
     console.log("error");
     console.log(err);
-    alert("ERROR : can't find the back");
+    alert();
   });
 
-function productData(sofa = "") {
+function sofa(sofa = "") {
   const { imageUrl, altTxt, name, colors, price, description } = sofa;
-  productPrice = price;
-  imgUrl = imageUrl;
-  altText = altTxt;
+  //productPrice = price;
+  //imgUrl = imageUrl;
+  //altText = altTxt;
   makeImage(imageUrl, altTxt);
   makeTitle(name);
   makePrice(price);
@@ -58,34 +58,55 @@ function makeColors(colors = "") {
   }
 }
 
+
+// ADD PRODUCTS TO CART
+
 const button = document.querySelector("#addToCart");
-button.addEventListener("click", () => {
+button.addEventListener("click", addToCart);
+
+function addToCart() {
+
   const color = document.querySelector("#colors").value;
   const quantity = document.querySelector("#quantity").value;
-  if (color == null || color === "" || quantity == null || quantity == 0) {
-    alert("Selectionnez une couleur/quantitée");
-  }
-  console.log("added to cart");
-
   const productInfo = {
     id: productId,
-    color: color,
-    quantity: quantity,
+    colors: color,
+    quantities: Number(quantity),
   };
 
+  if (color == null || color === "" || quantity == null || quantity == 0) {
+    alert("Selectionnez une couleur/quantité");
+    return true;
+  }
+  if (window.confirm(`Vous rendre au panier ?`))
+  window.location.href = "cart.html";
+
   let save = JSON.parse(localStorage.getItem("cart"));
-  console.log(save);
+  //console.log(save);
 
   if (save) {
-    save.push(productInfo);
-    localStorage.setItem("cart", JSON.stringify(save));
-    console.log(save);
+    const resultFind = save.find(
+      (el) => el.id === productId && el.colors === color
+    );
+    
+    if (resultFind) {
+      let newQuantite =
+        parseInt(productInfo.quantities) + parseInt(resultFind.quantities);
+      resultFind.quantities = newQuantite;
+      localStorage.setItem("cart", JSON.stringify(save));
+      //console.log(save);
+    } else {
+      save.push(productInfo);
+      localStorage.setItem("cart", JSON.stringify(save));
+      console.log(save);
+    }
   } else {
     save = [];
     save.push(productInfo);
     localStorage.setItem("cart", JSON.stringify(save));
-    console.log(save);
+    //console.log(save);
   }
-});
+  
+}
